+++ OOP2/code/src/lt/astrauskas/oop2/Player.scala
@@ -2,7 +2,7 @@ package lt.astrauskas.oop2
 
 class Player extends IPlayer {
 
-  def pointsImpl(): Int = 0
+  def pointsImpl(): Double = 0
   def statusImpl(): String = "Player status: " + points()
   def toStringImpl(): String = "Player"
